httpd: run as userspace kernel process with blocking accept loop
- Add accept queue to tcp_listener in net_internal.h (ACCEPT_QUEUE_SIZE=8)
- net_tcp.c: init queue on listen, enqueue conns when no callbacks set,
  add net_tcp_accept() with scheduler_yield() inside wait loop,
  reset rxlen/rx_fin on conn alloc, drain net_poll() after close
- net.h: declare net_tcp_accept()
- httpd.c: rewrite with httpd_task() loop (accept->handle->close),
  httpd_start/stop only manage listener and httpd_running flag
- httpd.h: add httpd_task() declaration
- kernel.c: process_create(httpd_task, 'httpd') alongside netd
- syscall.h: add SYS_NET_TCP_LISTEN..SYS_NET_TCP_UNLISTEN (183-188)
- syscall.c: implement 6 new TCP server syscall handlers
- process.c: grant TCP server syscalls to user default cap profile

All 272 e2e tests pass.

+++ src/shell/cmds/cmd_expr.c
@@ -21,7 +21,7 @@ void cmd_expr(const char *args) {
     long a = parse_num(&p);
     while (*p == ' ') p++;
 
-    if (!*p) { kprintf("%ld\n", (uint64_t)a); return; }
+    if (!*p) { kprintf("%d\n", (uint64_t)(int)a); return; }
 
     char op = *p++;
     /* Handle multi-char ops: !=, >=, <= */
@@ -54,5 +54,5 @@ void cmd_expr(const char *args) {
         return;
     }
 
-    kprintf("%ld\n", (uint64_t)result);
+    kprintf("%d\n", (uint64_t)(int)result);
 }
